rename project names
